requirements.txt
Flask
python-dotenv
ollama

